Function to ask a simple yes no question

Provides a simple wrapper around utils::askYesNo(). This function is used to ask yes no questions in an interactive way. It repeats the question until a valid answer is given.

Usage

ask_yes_no(msg, default = TRUE, prompts = c("Yes", "No", "Cancel"), ...)

Arguments

msg

The prompt message for the user.

default

The default response.

prompts

Any of: a character vector containing 3 prompts corresponding to return values of TRUE, FALSE, or NA, or a single character value containing the prompts separated by / characters, or a function to call.

...

Additional parameters, ignored by the default function.

See also