Cache organisation list from URL

This function retrieves the organisation list from a given URL and caches it locally. If the URL is "https://github.com/inbo", it uses the default INBO organisation list. For other URLs, it checks if a public citeme repository exists and clones it to retrieve the organisation list. The cached organisation list is stored in a specified configuration folder.

Usage

cache_org(url, config_folder = R_user_dir("citeme", "config"))

Arguments

url

The URL of the organisation list to retrieve.

config_folder

The folder where the cached organisation list should be stored. Defaults to the user's R configuration directory for citeme.

Value

The retrieved organisation list, or NULL if the URL is invalid or the repository.

See also